requirements.txt
Django==1.4.1
PIL==1.1.7
South==0.7.6
Werkzeug==0.8.3
amqplib==1.0.2
anyjson==0.3.3
barrel==0.1.3
billiard==2.7.3.18
boto==2.6.0
celery==3.0.12
distribute==0.6.28
dj-database-url==0.2.1
django-celery==3.0.11
django-debug-toolbar==0.9.4
django-extensions==0.9
django-grappelli==2.4.2
django-storages==1.1.5
django-tastypie==0.9.11
feedparser==5.1.2
gunicorn==0.14.6
kombu==2.4.8
mimeparse==0.1.3
psycopg2==2.4.5
python-dateutil==1.5
pytz==2012h
wsgiref==0.1.2
